+++ src/main/kotlin/race/model/FixedMoveCondition.kt
@@ -1,4 +1,4 @@
-package race
+package race.model
 
 class FixedMoveCondition(private val canGoStraight: Boolean) : MoveCondition {
     override fun canGoStraight(): Boolean {
