+++ src/main/kotlin/race/model/MoveCondition.kt
@@ -1,4 +1,4 @@
-package race
+package race.model
 
 interface MoveCondition {
     fun canGoStraight(): Boolean
