+++ src/main/kotlin/race/model/Race.kt
@@ -1,4 +1,4 @@
-package race
+package race.model
 
 import race.view.RaceResult
