+++ src/main/kotlin/race/model/RandomMoveCondition.kt
@@ -1,4 +1,4 @@
-package race
+package race.model
 
 import kotlin.random.Random
