+++ src/test/kotlin/race/CarTest.kt
@@ -3,6 +3,9 @@ package race
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
+import race.model.Car
+import race.model.FixedMoveCondition
+import race.model.RandomMoveCondition
 
 class CarTest {
     @Test
